+++ unta_food/service/pkg/controller/utna_food_controller.go
@@ -7,20 +7,8 @@ import (
 
 	"github.com/aws/aws-lambda-go/events"
 	"github.com/tsutarou10/line_project/service/pkg/entity"
-	"github.com/tsutarou10/line_project/service/pkg/usecase"
 	"github.com/tsutarou10/line_project/service/pkg/utils"
 )
-
-type Controller struct {
-	in usecase.UTNAFoodInputPort
-}
-
-func NewController(in usecase.UTNAFoodInputPort) *Controller {
-	log.Printf("[START] :%s", utils.GetFuncName())
-	defer log.Printf("[END] :%s", utils.GetFuncName())
-
-	return &Controller{in}
-}
 
 func (c *Controller) RegisterController(ctx context.Context, req events.APIGatewayProxyRequest) error {
 	log.Printf("[START] :%s", utils.GetFuncName())
@@ -39,23 +27,18 @@ func (c *Controller) RegisterController(ctx context.Context, req events.APIGatew
 		return errors.New(msg)
 	}
 
-	if !isURL(wc.ReceivedMessages[0]) {
+	if !utils.IsURL(wc.ReceivedMessages[0]) {
 		msg := "invalid url"
 		log.Printf("[ERROR]: %s, %s is %s", utils.GetFuncName(), wc.ReceivedMessages[0], msg)
 		return errors.New(msg)
 	}
+
 	input := entity.UTNAEntityFood{
 		URL: wc.ReceivedMessages[0],
 	}
+
 	if len(wc.ReceivedMessages) > 1 {
-		memo := ""
-		for i, rm := range wc.ReceivedMessages[1:] {
-			memo += rm
-			if i != len(wc.ReceivedMessages)-1 {
-				memo += " "
-			}
-		}
-		input.Memo = memo
+		input.Memo = createMemo(wc.ReceivedMessages[1:])
 	}
 	return c.in.HandleRegister(ctx, input)
 }
@@ -105,7 +88,7 @@ func (c *Controller) UpdateController(ctx context.Context, req events.APIGateway
 		return errors.New(msg)
 	}
 
-	if !isURL(wc.ReceivedMessages[2]) {
+	if !utils.IsURL(wc.ReceivedMessages[2]) {
 		msg := "invalid url"
 		log.Printf("[ERROR]: %s, %s is %s", utils.GetFuncName(), wc.ReceivedMessages[0], msg)
 		return errors.New(msg)
@@ -114,14 +97,7 @@ func (c *Controller) UpdateController(ctx context.Context, req events.APIGateway
 		URL: wc.ReceivedMessages[2],
 	}
 	if len(wc.ReceivedMessages) > 3 {
-		memo := ""
-		for i, rm := range wc.ReceivedMessages[3:] {
-			memo += rm
-			if i != len(wc.ReceivedMessages)-1 {
-				memo += " "
-			}
-		}
-		input.Memo = memo
+		input.Memo = createMemo(wc.ReceivedMessages[3:])
 	}
 	return c.in.HandleUpdate(ctx, input)
 }
